defined-4 fetch регистрации пользователя
